+++ Runtime/Stages/Mipmapper.cs
@@ -1,3 +1,4 @@
+using System.Collections.ObjectModel;
 using UnityEngine;
 using UnityEngine.Rendering;
 using UnityEngine.Experimental.Rendering;
@@ -22,6 +23,8 @@ public class Mipmapper {
   int _propSrc;
   CommandBuffer _command;
   ComputeShader _compute;
+  NumThreads _threadsFilter;
+  NumThreads _threadsShift;
   VXGI _vxgi;
 
   public Mipmapper(VXGI vxgi) {
@@ -29,13 +32,7 @@ public class Mipmapper {
 
     _command = new CommandBuffer { name = "VXGI.Mipmapper" };
 
-    if (Application.platform == RuntimePlatform.LinuxEditor || Application.platform == RuntimePlatform.LinuxPlayer) {
-      _kernelFilter = 1;
-    } else {
-      _kernelFilter = 0;
-    }
-
-    _kernelShift = compute.FindKernel("CSShift");
+    InitializeKernel();
 
     _propDisplacement = Shader.PropertyToID("Displacement");
     _propDst = Shader.PropertyToID("Dst");
@@ -48,18 +45,23 @@ public class Mipmapper {
   }
 
   public void Filter(ScriptableRenderContext renderContext) {
+    UpdateKernel();
+
     var radiances = _vxgi.radiances;
 
     for (var i = 1; i < radiances.Length; i++) {
       int resolution = radiances[i].volumeDepth;
-      int groups = Mathf.CeilToInt((float)resolution / 4f);
 
-      _command.BeginSample(_sampleFilter + resolution.ToString());
+      _command.BeginSample(_sampleFilter + resolution.ToString("D3"));
       _command.SetComputeIntParam(compute, _propDstRes, resolution);
       _command.SetComputeTextureParam(compute, _kernelFilter, _propDst, radiances[i]);
       _command.SetComputeTextureParam(compute, _kernelFilter, _propSrc, radiances[i - 1]);
-      _command.DispatchCompute(compute, _kernelFilter, groups, groups, groups);
-      _command.EndSample(_sampleFilter + resolution.ToString());
+      _command.DispatchCompute(compute, _kernelFilter,
+         Mathf.CeilToInt((float)resolution /_threadsFilter.x),
+         Mathf.CeilToInt((float)resolution /_threadsFilter.y),
+         Mathf.CeilToInt((float)resolution /_threadsFilter.z)
+      );
+      _command.EndSample(_sampleFilter + resolution.ToString("D3"));
     }
 
     renderContext.ExecuteCommandBuffer(_command);
@@ -67,17 +69,42 @@ public class Mipmapper {
   }
 
   public void Shift(ScriptableRenderContext renderContext, Vector3Int displacement) {
-    int groups = Mathf.CeilToInt((float)_vxgi.resolution / 4f);
+    UpdateKernel();
 
     _command.BeginSample(_sampleShift);
     _command.SetComputeIntParam(compute, _propDstRes, (int)_vxgi.resolution);
     _command.SetComputeIntParams(compute, _propDisplacement, new[] { displacement.x, displacement.y, displacement.z });
     _command.SetComputeTextureParam(compute, _kernelShift, _propDst, _vxgi.radiances[0]);
-    _command.DispatchCompute(compute, _kernelShift, groups, groups, groups);
+    _command.DispatchCompute(compute, _kernelShift,
+      Mathf.CeilToInt((float)_vxgi.resolution / _threadsShift.x),
+      Mathf.CeilToInt((float)_vxgi.resolution / _threadsShift.y),
+      Mathf.CeilToInt((float)_vxgi.resolution / _threadsShift.z)
+    );
     _command.EndSample(_sampleShift);
     renderContext.ExecuteCommandBuffer(_command);
     _command.Clear();
 
     Filter(renderContext);
+  }
+
+  void InitializeKernel() {
+    if (_vxgi.resolutionPlusOne) {
+      _kernelFilter = 0;
+    } else {
+      _kernelFilter = 2;
+    }
+
+    if (!VXGIRenderPipeline.isD3D11Supported) {
+      _kernelFilter += 1;
+    }
+
+    _kernelShift = compute.FindKernel("CSShift");
+    _threadsFilter = new NumThreads(compute, _kernelFilter);
+    _threadsShift = new NumThreads(compute, _kernelShift);
+  }
+
+  [System.Diagnostics.Conditional("UNITY_EDITOR")]
+  void UpdateKernel() {
+    InitializeKernel();
   }
 }